+++ pkg/ui/console/client.go
@@ -6,6 +6,7 @@ import (
 	"errors"
 	"fmt"
 	"os"
+	"path"
 	"strings"
 	"time"
 
@@ -56,13 +57,16 @@ func CreateClientFromConfig(ctx context.Context, cfg *config.Config, opts ...cli
 		c.PolicyDirectory = cfg.CloudQuery.PolicyDirectory
 		c.DSN = cfg.CloudQuery.Connection.DSN
 		c.SkipBuildTables = viper.GetBool("skip-build-tables")
+		c.HistoryCfg = cfg.CloudQuery.History
 	})
 	c, err := client.New(ctx, opts...)
 	if err != nil {
 		ui.ColorizedOutput(ui.ColorError, "❌ Failed to initialize client. Error: %s\n\n", err)
 		return nil, err
 	}
-	return &Client{c, cfg, progressUpdater}, err
+	client := &Client{c, cfg, progressUpdater}
+	client.checkForUpdate(ctx)
+	return client, err
 }
 
 func (c Client) DownloadProviders(ctx context.Context) error {
@@ -79,6 +83,10 @@ func (c Client) DownloadProviders(ctx context.Context) error {
 		c.updater.Wait()
 	}
 	ui.ColorizedOutput(ui.ColorProgress, "Finished provider initialization...\n\n")
+	updates, _ := c.c.CheckForProviderUpdates(ctx)
+	for _, u := range updates {
+		ui.ColorizedOutput(ui.ColorInfo, fmt.Sprintf("Update available for provider %s: %s ➡️ %s\n\n", u.Name, u.Version, u.LatestVersion))
+	}
 	return nil
 }
 
@@ -149,16 +157,17 @@ func (c Client) DownloadPolicy(ctx context.Context, args []string) error {
 	return nil
 }
 
-func (c Client) RunPolicies(ctx context.Context, args []string, policyName, outputDir, subPath string, stopOnFailure, skipVersioning, failOnViolation, noResults bool) error {
-	c.c.Logger.Debug("Received params: args: %v, policyName: %s, outputDir: %s, stopOnFailure: %v, skipVersioning: %v, failOnViolation: %v, noResults: %v", args, policyName, outputDir, stopOnFailure, skipVersioning, failOnViolation, noResults)
-
-	policiesToRun, err := client.FilterPolicies(args, c.cfg.Policies, policyName, subPath)
-
+func (c Client) RunPolicies(ctx context.Context, args []string, policyName, outputDir string, stopOnFailure, skipVersioning, failOnViolation, noResults bool) error {
+	c.c.Logger.Debug("Received params:", "args", args, "policyName", policyName, "outputDir", outputDir, "stopOnFailure", stopOnFailure, "skipVersioning", skipVersioning, "failOnViolation", failOnViolation, "noResults", noResults)
+	if err := c.DownloadProviders(ctx); err != nil {
+		return err
+	}
+	policiesToRun, err := policy.FilterPolicies(args, c.cfg.Policies, policyName)
 	if err != nil {
 		ui.ColorizedOutput(ui.ColorError, err.Error())
 		return err
 	}
-	c.c.Logger.Info("Policies to run: %v", policiesToRun)
+	c.c.Logger.Info("Policies to run", "policies", policiesToRun)
 
 	ui.ColorizedOutput(ui.ColorProgress, "Starting policies run...\n\n")
 
@@ -200,6 +209,48 @@ func (c Client) RunPolicies(ctx context.Context, args []string, policyName, outp
 
 	ui.ColorizedOutput(ui.ColorProgress, "Finished policies run...\n\n")
 	return nil
+}
+
+func (c Client) DescribePolicies(ctx context.Context, args []string, policyName string, skipVersioning bool) error {
+	policiesToDescribe, err := policy.FilterPolicies(args, c.cfg.Policies, policyName)
+	if err != nil {
+		ui.ColorizedOutput(ui.ColorError, err.Error())
+		return err
+	}
+	req := &client.PoliciesRunRequest{
+		Policies:       policiesToDescribe,
+		PolicyName:     policyName,
+		SkipVersioning: skipVersioning,
+	}
+	return c.describePolicies(ctx, req)
+}
+
+func (c Client) describePolicies(ctx context.Context, req *client.PoliciesRunRequest) error {
+	policies, err := c.c.LoadPolicies(ctx, req)
+	if err != nil {
+		ui.ColorizedOutput(ui.ColorError, err.Error())
+		return fmt.Errorf("failed to load policies: %w", err)
+	}
+	ui.ColorizedOutput(ui.ColorHeader, "Describe Policy %s output:\n\n", req.PolicyName)
+	t := &Table{writer: tablewriter.NewWriter(os.Stdout)}
+	t.SetHeaders("Path", "Description")
+	buildDescribePolicyTable(t, policies, "")
+	t.Render()
+	ui.ColorizedOutput(ui.ColorInfo, "To execute any policy use the path defined in the table above.\nFor example `cloudquery policy run %s %s`", req.PolicyName, getNestedPolicyExample(policies[0], ""))
+	return nil
+}
+
+func getNestedPolicyExample(p *policy.Policy, policyPath string) string {
+	if len(p.Policies) > 0 {
+		return getNestedPolicyExample(p.Policies[0], path.Join(policyPath, strings.ToLower(p.Name)))
+	}
+	return policyPath
+}
+func buildDescribePolicyTable(t ui.Table, pp policy.Policies, policyPath string) {
+	for _, p := range pp {
+		t.Append(path.Join(policyPath, strings.ToLower(p.Name)), p.Description)
+		buildDescribePolicyTable(t, p.Policies, path.Join(policyPath, strings.ToLower(p.Name)))
+	}
 }
 
 func (c Client) CallModule(ctx context.Context, req ModuleCallRequest) error {
@@ -298,7 +349,6 @@ func (c Client) UpgradeProviders(ctx context.Context, args []string) error {
 			return err
 		} else {
 			ui.ColorizedOutput(ui.ColorSuccess, "✓ Upgraded provider %s to %s successfully.\n\n", p.Name, p.Version)
-			color.GreenString("✓")
 		}
 	}
 	ui.ColorizedOutput(ui.ColorProgress, "Finished upgrading providers...\n\n")
@@ -422,6 +472,18 @@ func (c Client) getModuleProviders(ctx context.Context) ([]*cqproto.GetProviderS
 	}
 
 	return list, nil
+}
+
+func (c Client) checkForUpdate(ctx context.Context) {
+	v, err := client.MaybeCheckForUpdate(ctx, afero.Afero{Fs: afero.NewOsFs()}, time.Now().Unix(), client.UpdateCheckPeriod)
+	if err != nil {
+		c.c.Logger.Warn("update check failed", "error", err)
+		return
+	}
+	if v != nil {
+		ui.ColorizedOutput(ui.ColorInfo, "An update to CloudQuery core is available: %s!\n\n", v)
+	}
+	c.c.Logger.Debug("update check succeeded", "new_version", v.String())
 }
 
 func buildFetchProgress(ctx context.Context, providers []*config.Provider) (*Progress, client.FetchUpdateCallback) {
@@ -612,7 +674,7 @@ func defineResultColumnWidths(execResult []*policy.QueryResult) string {
 }
 
 func findOutput(columnNames []string, data [][]interface{}) []string {
-	outputKeys := []string{"id", "identifier", "resource_idnetifier", "uid", "uuid", "arn"}
+	outputKeys := []string{"id", "identifier", "resource_identifier", "uid", "uuid", "arn"}
 	outputKey := ""
 	outputResources := make([]string, 0)
 	for _, key := range outputKeys {